RVA_020: Drop M-mode extension Smcntrpmf
The Server Platform specification are written to specify requirements
as seen by S-mode and lower. The rationale is to not requiring
explicit implementation of M-mode ISA.

The features provided by Smcntrpmf are implied if Ssccfg is
implemented. So it is safe to drop the reference as long as Ssccfg is
specififed.

Considering the above two points, drop the reference to Smcntrpmf.

Signed-off-by: Punit Agrawal <punit.agrawal@bytedance.com>