FIRM: Require RIMT for IOMMU
While BRS does have requirements dictating which tables to use, like
ACPI_030, conditional requirements like "If there's an IOMMU, then
the RIMT must be used" would require the BRS to consider what there
might be. The platform server spec is better positioned for that
since it's sure there may be an IOMMU, as it also builds on the
server-soc spec. So, as there's some concern that a RIMT won't be
used when there is an IOMMU, require firmware to use it.

Resolves: https://github.com/riscv-non-isa/riscv-server-platform/issues/2
Signed-off-by: Andrew Jones <ajones@ventanamicro.com>