Drop HPER_070
We require BRS-I which has URT_010 and states systems without an RTC
must implement EFI GetTime(). BRS-I also has URT_020 which states that
it's expected for the OS to use the RTC directly on systems with one
on an OS-managed bus.

So the BRS has us covered already, drop the RTC requirement.

Resolves: https://github.com/riscv-non-isa/riscv-server-platform/issues/54
Signed-off-by: Andrew Jones <ajones@ventanamicro.com>