FIRM: simplify calls to UEFI runtime services
The worst case, without FIRM_120, is that the OS would have to toggle
landing pads (through SBI FWFT) around every call to a runtime service.

FIRM_130 is stating the obvious in case software sanity-checks the
server platform before enabling landing pads.

The compilers currently generate a `lpad 0` at the head of each
function, so FIRM_140 ensures that an OS can depend on this behavior,
and call the runtime service without fallback paths.

Signed-off-by: Radim Krčmář <rkrcmar@ventanamicro.com>