Don't reference domains
The purpose of commit 7e5b273ffbe7 ("Constrain RVA_030 and RVA_040
to apply internally to each supervisor domain") is to point out that
a platform need not be globally homogeneous as long as an OS is
constrained to a homogeneous environment. While domains are one way
of providing those environments, we should only call out "environment",
not examples of environment providers.

Fixes: 7e5b273ffbe7 ("Constrain RVA_030 and RVA_040 to apply internally to each supervisor domain")
Signed-off-by: Andrew Jones <ajones@ventanamicro.com>