Attempt to resolve security issues and conflicts between SEC_010, SEC_011, SEC_060
Resolve the conflicts between SEC_010, SEC_011, and SEC_060, while also
qualifying the requirement that previously mandated that physically
present users must be able to bypass system security.

Regarding the "physically present users" issue: during the discussion
on how to resolve this topic, Andrea Bolognani of Red Hat suggested
referring to the _Windows Hardware Compatibility Specifications for
Windows 11, version 24H2 - Systems_ document, available here:

https://go.microsoft.com/fwlink/?linkid=2272281

The authors of this document qualify their requirements here by only
making them applicable to "systems that are [not] intended to be locked
down".  On the theory that if it's good enough for Microsoft Windows
and Red Hat, it's good enough for the RISC-V Server Platform specification,
adopt this qualification for these requirements.

[Text edited by Ved.]

Resolves: https://github.com/riscv-non-isa/riscv-server-platform/issues/44
Resolves: https://github.com/riscv-non-isa/riscv-server-platform/issues/45