Platform requirements: add the battery-backed RTC requirement back
After discussion in the 2025-07-07 RISC-V Server Platform meeting,
resuscitate the HPER_070 requirement for a battery-backed Real Time
Clock (RTC), but with some additional context to describe how it should
be used.  This "Server Platform RTC" is primarily intended to support
time-based certificate validation for UEFI security applications, and
may be separate from other RTCs that might be dedicated to operating
systems or security elements.  Read access to this Server Platform RTC
may also be provided by platform firmware to operating systems via the
EFI_GET_TIME mechanism to provide an initial date and time source.
Write (clock set) access may optionally be provided by platform
firmware to operating systems in cases where there are no security
requirements that forbid it.  Clarify that this is a risk not to the
OS itself, but to higher-privilege firmware, or to other entities that
use the RTC that are in separate privilege domains.

Closes #74.
Closes #38.

Cc: Ved Shanbogue <ved@rivosinc.com>
Cc: Carl Perry <caperry@edolnx.net>
Cc: Andrea Bolognani <abologna@redhat.com>
Cc: Andrei Warkentin <andrei.warkentin@intel.com>