Requirements: renumber to ensure sufficient space between each requirement
Renumber several of the requirements, and update corresponding IDs in the
text, to ensure equal spacing between each requirement.  The goal here
is to facilitate future additions as may be necessary.

Ved was kind enough to hold off renumbering requirements in a previous
pull request until we reached a stable version.  This was because many
requirements were under discussion and it seemed too fraught with
difficulty to have multiple requirement IDs referring to the same
requirement (pre- and post- renumbering).

It's unclear to me whether now, as we approach stable, is the right
time; or whether we should wait until after public review.

Cc: Ved Shanbogue <ved@rivosinc.com>