RISC-V option ROMs only should be supported
Depending on the firmware, there may be alternative ways to provide
drivers. As there could be a security argument for disabling option
ROM support and there's the fact that no RISC-V option ROMs exist at
this time, then make this rule a SHOULD instead of a MUST. The rule
now needs to be split to ensure that authentication support is not
interpreted as a should, since, if options ROMs are supported, then
authentication support MUST be supported.

Resolves: https://github.com/riscv-non-isa/riscv-server-platform/issues/81
Signed-off-by: Andrew Jones <ajones@ventanamicro.com>