RVA_050: clarify that identifiers can be used to distinguish the harts
As agreed in the server platform meeting, the current wording is too
strong, because the harts are expected to differ in product identifiers.

The harts must also differ in hart id, but that one is not explicitly
called as the architecture would not be RISC-V otherwise, and it's also
not directly accessible through SBI ecall (it is getting passed in a
register at boot).

The spirit of the rule is to make sure that the OS doesn't have to jump
through hoops to find a reasonable configuration that would work across
all harts.

Co-authored-by: Vedvyas Shanbhogue <ved@rivosinc.com>
Signed-off-by: Radim Krčmář <rkrcmar@ventanamicro.com>