ME_RVA_010_010: do not assume hart 0 is exposed via RHTC
The server platform spec requires that all harts exposed to the operating
systems support the same set of extensions. But there is no need for the
first hart exposed to the operating system to be numbered 0.

E.g. hart 0 may be a management hart which is not exposed to the OS.

Signed-off-by: Heinrich Schuchardt <heinrich.schuchardt@canonical.com>