Reduce extension mandates
Based on RVI Architecture Review Committee (ARC) feedback[1] that
clarifies the servers this server platform spec have in mind are not
limited to "datacenter servers" nor "enterprise servers", demote a
couple previously mandated extensions to optional. ARC states that
there's an expectation for the spec to mandate them in later
versions, however, so don't just drop them from the spec (as any
optional extension is optional whether mentioned in the spec or
not). Instead, state the extensions SHOULD be supported and, in
non-normative text, that they may be mandated in later versions of
the spec.

Link: https://lists.riscv.org/g/tech-privileged/message/2622 [1]
Signed-off-by: Andrew Jones <ajones@ventanamicro.com>