tests: note validation of referenced specifications, and clean up
Note that the Server Platform Test Specification will include some
tests of referenced specifications, using some features specified
in the Server Platform specification, and/or features specified in
specifications that are included in the Server Platform
specification by reference.  These tests use a "C" character for
the test type (denoting that they are composite tests).

Remove a comment that most tests simply check for the existence of
a given feature.  While it may be true, this comment seems
superfluous, and may rapidly become obsolete.

Clean up some whitespace while here.