RVA_060/070: Move to new SEE section
This is a change suggested in ARC feedback.

From Ved:
"""
The debug related registers to setup triggers are not visible to
ISA. The RVA_060 and RVA_070 are discussing that. The Sdext and
Sdtrig - the parts of debug that are visible to S-mode - are
included in RVA020.

I would suggest a section called "RISC-V Hart SEE" and place the
RVA_060 and RVA_070 in that section.
"""

Signed-off-by: Andrew Jones <ajones@ventanamicro.com>